refactor(storage): extract seal_store + org_store from db.rs (lock-critical, verbatim)
Pure MOVE, zero behavior change. The at-rest SEAL machinery (21 methods)
and the ORG / shared-brain DB methods (39) lift out of storage/db.rs into
storage/seal_store.rs + storage/org_store.rs; db.rs 20,443 -> 18,484 (-1,959).

All 10 moved blocks (2,087 lines) are present VERBATIM (contiguous
byte-identical runs) in the new files. NO seal/blank/gate/blob body changed.
The only line-level edits in the whole diff are visibility widenings + wiring:
- 11 shared derived-table purge choke-points -> pub(crate) (each forced by a
  seal_store cross-file Self:: call in blank_sealed_notes_in_folders /
  discard_folder_seal / reblank_locked_folders_at_rest); KEPT in db.rs since
  they are shared across delete/seal/relock.
- fts_match_query -> pub(crate) (forced by org_store search_org_chunks_fts;
  still used by 4 retained db.rs FTS readers).
- removed a now-unused OrgChunkHit import.

Seal writers keep exact write ordering; the verify-before-destroy contract is
untouched (callers stay in commands, out of this diff); all 19
text_blob/data_blob IS NOT NULL re-blank guards verbatim. Gated org readers
keep context_enabled=1 + tombstoned=0 verbatim (a disabled/tombstoned org
item stays invisible).

Tests all stayed in db.rs mod tests: db.rs #[test] 250 = trunk; storage/ total
262 = trunk; pub fn conservation exact (db.rs 94 + seal 21 + org 39 = 154 =
trunk). Verified: cargo build --lib 0 warnings, clippy --lib -D warnings clean,
round_trips_byte_identical 11/11 (incl seal_transcript_timeline / seal_document
/ seal_manual_notes / lock_unlock), seal 145, org 139, blob 7, lock 175,
visible 31, lifecycle 313, storage::db 250 — all 0 failed. MSRV grep clean.
