+++ src-tauri/src/storage/org_store.rs
@@ -1,0 +1,1154 @@
+//! Shared-Brain / ORG storage surface — the local org membership state, the outbound org-share
+//! state machine (the `org_shares` table + its dedup/retry/revoke bookkeeping), and the decrypted
+//! ORG-ITEM replica (`org_items` + its `org_chunks` / `org_vec_chunks` / `fts_org_chunks` index,
+//! the KNN + FTS retrieval legs, and the read-only item/list getters). Extracted VERBATIM from
+//! `storage::db` (God-file split, a PURE MOVE — zero behavior change): an inherent-impl split of
+//! [`crate::storage::db::Db`] across files; every method keeps its EXACT prior body, signature, AND
+//! gating.
+//!
+//! GATING — org items are deliberately ORG-DISCLOSED content that lives OUTSIDE the per-folder lock
+//! domain (SQLCipher protects them at rest; no folder seal/`visibility_clause` gate applies — spec
+//! §"Trust model"). The read gate that DOES apply is the PER-INSTANCE org toggle
+//! `os.context_enabled = 1`, joined + filtered at the SQL level in `search_org_chunks_knn` /
+//! `search_org_chunks_fts` / `get_org_item` / `count_org_items` EXACTLY as on trunk — a disabled
+//! org's chunks/items are excluded in SQL, never read into Rust. `get_org_item` / `list_org_items`
+//! stay gated on `tombstoned = 0` (+ `context_enabled` where trunk had it). The org-private helpers
+//! (`map_org_state`, `map_org_share`, the `ORG_SHARE_COLS` const, `dedup_org_hits_by_item`, and
+//! `purge_org_item_chunks_tx`) are used ONLY by these methods, so they moved along and stay PRIVATE
+//! to this module. `fts_match_query` (shared with several db.rs FTS readers) was promoted to
+//! `pub(crate)` in `db.rs` and is reached cross-file. The 1:1-share (`outbound_shares`) machinery
+//! stays in `db.rs`. Tests stay in db.rs's `mod tests` (shared harness); the count is conserved.
+
+use std::collections::HashSet;
+
+use rusqlite::OptionalExtension;
+
+use crate::embed::Embedder;
+use crate::error::Result;
+use crate::storage::db::{fts_match_query, map_err, Db};
+use crate::storage::models::OrgChunkHit;
+
+impl Db {
+    /// Upsert the locally-cached membership of an org (create/status). Preserves an existing row's
+    /// local `consented` flag, `last_seq` cursor, AND `context_enabled` toggle (an incoming status
+    /// refresh MUST NOT reset the consent flag, rewind the sync cursor, or silently re-enable an org
+    /// the user disabled on this instance). NO content — membership metadata only.
+    pub fn upsert_org_state(&self, o: &crate::storage::OrgState) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "INSERT INTO org_state (org_id, name, role, joined_at, consented, last_seq, generation, context_enabled)
+               VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)
+             ON CONFLICT(org_id) DO UPDATE SET
+               name = excluded.name,
+               role = excluded.role,
+               generation = excluded.generation",
+            rusqlite::params![
+                o.org_id,
+                o.name,
+                o.role,
+                o.joined_at,
+                o.consented as i64,
+                o.last_seq,
+                o.generation as i64,
+                o.context_enabled as i64
+            ],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    fn map_org_state(r: &rusqlite::Row<'_>) -> rusqlite::Result<crate::storage::OrgState> {
+        Ok(crate::storage::OrgState {
+            org_id: r.get(0)?,
+            name: r.get(1)?,
+            role: r.get(2)?,
+            joined_at: r.get(3)?,
+            consented: r.get::<_, i64>(4)? != 0,
+            last_seq: r.get(5)?,
+            generation: r.get::<_, i64>(6)? as u32,
+            context_enabled: r.get::<_, i64>(7)? != 0,
+        })
+    }
+
+    /// The locally-cached state of one org (or `None` if not joined locally).
+    pub fn get_org_state(&self, org_id: &str) -> Result<Option<crate::storage::OrgState>> {
+        let conn = self.lock();
+        conn.query_row(
+            "SELECT org_id, name, role, joined_at, consented, last_seq, generation, context_enabled
+               FROM org_state WHERE org_id = ?1",
+            rusqlite::params![org_id],
+            Self::map_org_state,
+        )
+        .optional()
+        .map_err(map_err)
+    }
+
+    /// Every locally-joined org (for the launch sweep + a future multi-org list). Ordered by join time.
+    pub fn list_org_states(&self) -> Result<Vec<crate::storage::OrgState>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT org_id, name, role, joined_at, consented, last_seq, generation, context_enabled
+                   FROM org_state ORDER BY joined_at ASC",
+            )
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map([], Self::map_org_state)
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// Set the local org-egress consent flag for an org (mirrors the config consent grants: the ONLY
+    /// mutator, so a status refresh can't clear it). Fail-safe ordering is the caller's concern.
+    pub fn set_org_consented(&self, org_id: &str, consented: bool) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_state SET consented = ?2 WHERE org_id = ?1",
+            rusqlite::params![org_id, consented as i64],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Set the PER-INSTANCE org context toggle (Settings → Organization): whether a JOINED org
+    /// contributes content on THIS Murmur install — browsing (`list_org_items`) AND brain/assistant
+    /// context (`search_org_chunks_knn`/`_fts`). The ONLY mutator (mirrors `set_org_consented`) — a
+    /// status/feed refresh (`upsert_org_state`) never touches this column. Disabling never deletes the
+    /// local replica; re-enabling is instant with no re-sync.
+    pub fn set_org_context_enabled(&self, org_id: &str, enabled: bool) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_state SET context_enabled = ?2 WHERE org_id = ?1",
+            rusqlite::params![org_id, enabled as i64],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Advance the synced feed cursor for an org (monotonic; a caller never rewinds it below the
+    /// stored value). Used by the feed-sync slice; kept here so the schema owner defines the writer.
+    pub fn set_org_last_seq(&self, org_id: &str, last_seq: i64) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_state SET last_seq = ?2 WHERE org_id = ?1 AND ?2 > last_seq",
+            rusqlite::params![org_id, last_seq],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Update the cached live generation for an org (after a rotation the owner drove, or a status pull).
+    pub fn set_org_generation(&self, org_id: &str, generation: u32) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_state SET generation = ?2 WHERE org_id = ?1",
+            rusqlite::params![org_id, generation as i64],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Drop the local org row (on leave / removal). Idempotent; leaves `org_shares` alone (a leave
+    /// doesn't retroactively un-share — the items stay published unless explicitly revoked).
+    pub fn delete_org_state(&self, org_id: &str) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "DELETE FROM org_state WHERE org_id = ?1",
+            rusqlite::params![org_id],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Insert a fresh outbound org share in the `queued` state (the "Share to Brain" action). The
+    /// caller sets `meeting_id` XOR `document_id`. `content_sha256` is the plaintext-envelope hash.
+    #[allow(clippy::too_many_arguments)]
+    pub fn insert_org_share(
+        &self,
+        id: &str,
+        org_id: &str,
+        meeting_id: Option<&str>,
+        document_id: Option<&str>,
+        kind: &str,
+        title: Option<&str>,
+        rev: u32,
+        generation: u32,
+        content_sha256: &[u8],
+        created_at: &str,
+    ) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "INSERT INTO org_shares
+               (id, org_id, meeting_id, document_id, kind, title, rev, generation,
+                content_sha256, item_id, state, last_error, created_at, updated_at)
+             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, NULL, 'queued', NULL, ?10, ?10)",
+            rusqlite::params![
+                id,
+                org_id,
+                meeting_id,
+                document_id,
+                kind,
+                title,
+                rev as i64,
+                generation as i64,
+                content_sha256,
+                created_at
+            ],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Advance a queued org share to `uploaded`, recording the server-assigned `item_id`. Clears any
+    /// prior error. Idempotent on the share id.
+    pub fn set_org_share_uploaded(
+        &self,
+        id: &str,
+        item_id: &str,
+        updated_at: &str,
+    ) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_shares SET state = 'uploaded', item_id = ?2, last_error = NULL,
+               updated_at = ?3 WHERE id = ?1",
+            rusqlite::params![id, item_id, updated_at],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Mark an org share `failed` with a non-PII error string (for the launch sweep to retry / the FE
+    /// to surface). The error is a fixed message + status, never note content.
+    pub fn set_org_share_failed(&self, id: &str, error: &str, updated_at: &str) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_shares SET state = 'failed', last_error = ?2, updated_at = ?3 WHERE id = ?1",
+            rusqlite::params![id, error, updated_at],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// Set an org share's state directly (e.g. `queued` → retry a `failed`, `uploaded` →
+    /// `revoke_pending`, `revoke_pending` → `revoked`). Idempotent; unknown id is a no-op.
+    pub fn set_org_share_state(&self, id: &str, state: &str, updated_at: &str) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_shares SET state = ?2, updated_at = ?3 WHERE id = ?1",
+            rusqlite::params![id, state, updated_at],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    fn map_org_share(r: &rusqlite::Row<'_>) -> rusqlite::Result<crate::storage::OrgShareRow> {
+        Ok(crate::storage::OrgShareRow {
+            id: r.get(0)?,
+            org_id: r.get(1)?,
+            meeting_id: r.get(2)?,
+            document_id: r.get(3)?,
+            kind: r.get(4)?,
+            title: r.get(5)?,
+            rev: r.get::<_, i64>(6)? as u32,
+            generation: r.get::<_, i64>(7)? as u32,
+            content_sha256: r.get(8)?,
+            item_id: r.get(9)?,
+            state: r.get(10)?,
+            last_error: r.get(11)?,
+            created_at: r.get(12)?,
+            updated_at: r.get(13)?,
+        })
+    }
+
+    const ORG_SHARE_COLS: &'static str =
+        "id, org_id, meeting_id, document_id, kind, title, rev, generation,
+         content_sha256, item_id, state, last_error, created_at, updated_at";
+
+    /// One org share by its local id.
+    pub fn get_org_share(&self, id: &str) -> Result<Option<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        conn.query_row(
+            &format!("SELECT {} FROM org_shares WHERE id = ?1", Self::ORG_SHARE_COLS),
+            rusqlite::params![id],
+            Self::map_org_share,
+        )
+        .optional()
+        .map_err(map_err)
+    }
+
+    /// The org share bearing a given server `item_id` (for revoke-by-item + self-share dedup).
+    pub fn org_share_by_item(&self, item_id: &str) -> Result<Option<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        conn.query_row(
+            &format!(
+                "SELECT {} FROM org_shares WHERE item_id = ?1",
+                Self::ORG_SHARE_COLS
+            ),
+            rusqlite::params![item_id],
+            Self::map_org_share,
+        )
+        .optional()
+        .map_err(map_err)
+    }
+
+    /// Every LIVE-OR-STUCK-LIVE org share anchored to a given source (`meeting_id` XOR
+    /// `document_id`). Powers the re-publish-on-edit fix: one logical note may be shared into SEVERAL
+    /// orgs, so this returns rows ACROSS ALL of them (never restricted to the first). Returns `uploaded`
+    /// rows AND `failed` rows that still carry a non-null `item_id` — the latter is a row whose MOST
+    /// RECENT republish attempt failed transiently (network blip during OCK acquire / seal / blob
+    /// upload / item publish) but whose PRIOR publish is still genuinely live on the server:
+    /// `set_org_share_failed` deliberately does not clear `item_id` on a republish failure (only the
+    /// SUCCESS path's `reset_org_share_for_retry` does), so such a row represents a live item whose
+    /// latest edit hasn't synced yet — not a dead row. Excluding it here (the pre-fix behavior) made it
+    /// permanently invisible to every caller keyed off this function (the edit-save republish path, the
+    /// re-share-block check, the Library share badge), so it could never self-heal and a manual re-share
+    /// would mint a genuine duplicate item. A `queued`/never-published `failed` row (no `item_id`, no
+    /// live server item to supersede yet — the launch sweep publishes the current plaintext for it) and
+    /// a `revoked`/`revoke_pending` share (intentionally torn down; an edit must not resurrect it) are
+    /// still excluded. Exactly one of `meeting_id`/`document_id` must be `Some`; both-`None` returns an
+    /// empty vec (no source ⇒ nothing to republish).
+    pub fn org_shares_for_source(
+        &self,
+        meeting_id: Option<&str>,
+        document_id: Option<&str>,
+    ) -> Result<Vec<crate::storage::OrgShareRow>> {
+        if meeting_id.is_none() && document_id.is_none() {
+            return Ok(Vec::new());
+        }
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(&format!(
+                "SELECT {} FROM org_shares
+                   WHERE (state = 'uploaded' OR (state = 'failed' AND item_id IS NOT NULL))
+                     AND ((?1 IS NOT NULL AND meeting_id = ?1)
+                       OR (?2 IS NOT NULL AND document_id = ?2))
+                   ORDER BY created_at ASC",
+                Self::ORG_SHARE_COLS
+            ))
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![meeting_id, document_id], Self::map_org_share)
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// Every LIVE (`uploaded`) org share of ONE exact source (`meeting_id` XOR `document_id`) in ONE
+    /// org, OLDEST-FIRST (stable tie-break on `id`). The `(org, source)`-scoped twin of
+    /// `org_shares_for_source` (which spans all orgs): powers the share IDEMPOTENCY guard + the
+    /// duplicate collapse — `[0]` is the canonical KEEPER (earliest published, the identity other
+    /// members first saw), `[1..]` are accidental duplicates to tombstone. `state = 'uploaded'` only
+    /// (a queued/failed row has no live server item; a revoked one was intentionally torn down).
+    /// `meeting_id`/`document_id` matched NULL-safe via `IS`; both-None ⇒ empty (no source).
+    pub fn uploaded_org_shares_for_source_in_org(
+        &self,
+        org_id: &str,
+        meeting_id: Option<&str>,
+        document_id: Option<&str>,
+    ) -> Result<Vec<crate::storage::OrgShareRow>> {
+        if meeting_id.is_none() && document_id.is_none() {
+            return Ok(Vec::new());
+        }
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(&format!(
+                "SELECT {} FROM org_shares
+                   WHERE org_id = ?1 AND state = 'uploaded'
+                     AND meeting_id IS ?2 AND document_id IS ?3
+                   ORDER BY created_at ASC, id ASC",
+                Self::ORG_SHARE_COLS
+            ))
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(
+                rusqlite::params![org_id, meeting_id, document_id],
+                Self::map_org_share,
+            )
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// Every DUPLICATE live org share across the whole DB: an `uploaded` row that has an EARLIER
+    /// `uploaded` sibling for the same `(org_id, meeting_id, document_id)` — i.e. the extras to
+    /// tombstone, keeping only the earliest per group. Powers the on-launch dedup sweep that cleans
+    /// duplicates created before the idempotency guard existed (e.g. a double-click on Share). Tie-break
+    /// on `id` so two rows sharing a `created_at` still pick ONE deterministic keeper. NEVER returns a
+    /// keeper (the earliest of its group). `meeting_id`/`document_id` grouped NULL-safe via `IS`.
+    pub fn duplicate_uploaded_org_shares(&self) -> Result<Vec<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(&format!(
+                "SELECT {} FROM org_shares o
+                   WHERE o.state = 'uploaded'
+                     AND EXISTS (
+                       SELECT 1 FROM org_shares e
+                        WHERE e.state = 'uploaded'
+                          AND e.org_id = o.org_id
+                          AND e.meeting_id IS o.meeting_id
+                          AND e.document_id IS o.document_id
+                          AND (e.created_at < o.created_at
+                            OR (e.created_at = o.created_at AND e.id < o.id)))
+                   ORDER BY o.created_at ASC, o.id ASC",
+                Self::ORG_SHARE_COLS
+            ))
+            .map_err(map_err)?;
+        let rows = stmt.query_map([], Self::map_org_share).map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// Cancel (mark `revoked`) any NOT-yet-uploaded (`queued`/`failed`) org share for a given
+    /// (org, source). Used after a collapse when a live `uploaded` keeper already exists for that
+    /// source: a pending sibling is redundant (the source is already live) and would otherwise linger
+    /// as a stuck "pending" row that the launch sweep re-attempts every start. These rows have NO server
+    /// `item_id`, so cancelling is LOCAL-ONLY (no tombstone). Returns the number of rows cancelled.
+    /// `meeting_id`/`document_id` matched NULL-safe via `IS`; both-None ⇒ 0 (no source).
+    pub fn cancel_pending_org_shares_for_source_in_org(
+        &self,
+        org_id: &str,
+        meeting_id: Option<&str>,
+        document_id: Option<&str>,
+        updated_at: &str,
+    ) -> Result<usize> {
+        if meeting_id.is_none() && document_id.is_none() {
+            return Ok(0);
+        }
+        let conn = self.lock();
+        let n = conn
+            .execute(
+                "UPDATE org_shares SET state = 'revoked', updated_at = ?4
+                   WHERE org_id = ?1 AND state IN ('queued', 'failed')
+                     AND meeting_id IS ?2 AND document_id IS ?3",
+                rusqlite::params![org_id, meeting_id, document_id, updated_at],
+            )
+            .map_err(map_err)?;
+        Ok(n)
+    }
+
+    /// SB-3 dedup: the EXISTING retriable (`queued`/`failed`) org-share row for a logical share key
+    /// (org + meeting-or-document), if any. `share_to_org_inner` REUSES it on a re-publish instead of
+    /// minting a fresh row every sweep tick — without this, each failed retry inserted a NEW row while
+    /// the old one survived, so a persistently-failing share amplified rows unboundedly and a later
+    /// recovery double-published. Newest-created first (a stable pick if somehow >1 exists). Uploaded/
+    /// revoked/revoke_pending rows are NOT reused (an uploaded share is a distinct published item; a
+    /// revoked one is intentionally torn down). `meeting_id`/`document_id` are matched exactly (both
+    /// NULL-safe via `IS`).
+    pub fn find_reusable_org_share(
+        &self,
+        org_id: &str,
+        meeting_id: Option<&str>,
+        document_id: Option<&str>,
+    ) -> Result<Option<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        conn.query_row(
+            &format!(
+                "SELECT {} FROM org_shares
+                   WHERE org_id = ?1
+                     AND meeting_id IS ?2 AND document_id IS ?3
+                     AND state IN ('queued', 'failed')
+                   ORDER BY created_at DESC LIMIT 1",
+                Self::ORG_SHARE_COLS
+            ),
+            rusqlite::params![org_id, meeting_id, document_id],
+            Self::map_org_share,
+        )
+        .optional()
+        .map_err(map_err)
+    }
+
+    /// SB-3 retry re-arm: reset an EXISTING org-share row back to `queued` for a fresh publish attempt,
+    /// refreshing the per-attempt fields (title/content hash/generation/timestamps) and CLEARING any
+    /// item_id + last_error. Used by `share_to_org_inner` when it reuses a `find_reusable_org_share`
+    /// row instead of inserting a new one — so N failed attempts stay ONE row, and a later success
+    /// flips that same row to uploaded (no duplicate). Idempotent on the row id.
+    pub fn reset_org_share_for_retry(
+        &self,
+        id: &str,
+        title: Option<&str>,
+        rev: u32,
+        generation: u32,
+        content_sha256: &[u8],
+        updated_at: &str,
+    ) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_shares SET state = 'queued', item_id = NULL, last_error = NULL,
+               title = ?2, rev = ?3, generation = ?4, content_sha256 = ?5, updated_at = ?6
+             WHERE id = ?1",
+            rusqlite::params![id, title, rev as i64, generation as i64, content_sha256, updated_at],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// All org shares in a given `state` (the launch sweep pulls `queued` + `revoke_pending`).
+    pub fn list_org_shares_in_state(
+        &self,
+        state: &str,
+    ) -> Result<Vec<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(&format!(
+                "SELECT {} FROM org_shares WHERE state = ?1 ORDER BY created_at ASC",
+                Self::ORG_SHARE_COLS
+            ))
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![state], Self::map_org_share)
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// Every LIVE org share across ALL sources/orgs — the un-scoped twin of `org_shares_for_source`,
+    /// for callers that need the whole "is this live" set rather than one source (the Library bulk
+    /// share-badge listing). Same STUCK-REPUBLISH definition of "live" as `org_shares_for_source`:
+    /// `uploaded` rows AND `failed` rows that still carry a non-null `item_id` — a row whose MOST
+    /// RECENT republish attempt failed transiently but whose PRIOR publish is still genuinely live on
+    /// the server (`set_org_share_failed` deliberately never clears `item_id`; only the success path's
+    /// `reset_org_share_for_retry` does). A `queued`/never-published `failed` row (no `item_id`) or a
+    /// `revoked`/`revoke_pending` share is excluded. See `org_shares_for_source` for the full rationale.
+    pub fn list_live_org_shares(&self) -> Result<Vec<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(&format!(
+                "SELECT {} FROM org_shares
+                   WHERE state = 'uploaded' OR (state = 'failed' AND item_id IS NOT NULL)
+                   ORDER BY created_at ASC",
+                Self::ORG_SHARE_COLS
+            ))
+            .map_err(map_err)?;
+        let rows = stmt.query_map([], Self::map_org_share).map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// Every org share for an org (the FE list). Newest first.
+    pub fn list_org_shares_for_org(
+        &self,
+        org_id: &str,
+    ) -> Result<Vec<crate::storage::OrgShareRow>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(&format!(
+                "SELECT {} FROM org_shares WHERE org_id = ?1 ORDER BY created_at DESC",
+                Self::ORG_SHARE_COLS
+            ))
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![org_id], Self::map_org_share)
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// The ACTIVE-OR-STUCK-LIVE (queued/uploaded/revoke_pending, PLUS a `failed` row that still
+    /// carries a non-null `item_id`) org shares anchored to a folder's meetings + notes, for the
+    /// lock×shares warn/revoke dialog. Content-free enough for the dialog (an `(item_id?, title?)`
+    /// pair per share; titles render only to the local owner).
+    ///
+    /// The `failed AND item_id IS NOT NULL` clause mirrors the definition of "live" established by
+    /// `org_shares_for_source`/`list_live_org_shares` (see their doc comments): `set_org_share_failed`
+    /// deliberately never clears `item_id` on a republish failure, so such a row's PRIOR publish is
+    /// still genuinely live on the server even though the row's state says `failed`. Before this fix
+    /// that shape was invisible to the lock×shares dialog and to bulk-revoke
+    /// (`active_org_share_ids_for_folder`), so locking a folder with a stuck failed-republish share
+    /// never warned the user and never tombstoned the still-live server item.
+    pub fn active_org_shares_for_folder(
+        &self,
+        folder_id: &str,
+    ) -> Result<Vec<(Option<String>, Option<String>)>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT s.item_id, s.title
+                   FROM org_shares s
+                   LEFT JOIN notes n  ON n.meeting_id = s.meeting_id
+                   LEFT JOIN documents d ON d.id = s.document_id
+                  WHERE (s.state IN ('queued','uploaded','revoke_pending')
+                     OR (s.state = 'failed' AND s.item_id IS NOT NULL))
+                    AND (n.folder_id = ?1 OR d.folder_id = ?1)",
+            )
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![folder_id], |r| {
+                Ok((r.get::<_, Option<String>>(0)?, r.get::<_, Option<String>>(1)?))
+            })
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// The folder's ACTIVE-OR-STUCK-LIVE org shares as `(row_id, item_id?, title)` for bulk-revoke:
+    /// an uploaded row (item_id present) is tombstoned server-side; a still-`queued` row (no item_id)
+    /// is cancelled locally so the launch sweep never egresses it; a `failed` row with a non-null
+    /// `item_id` (a republish attempt failed but the PRIOR publish is still live — see
+    /// [`Self::active_org_shares_for_folder`]) is tombstoned server-side same as an uploaded row.
+    /// Same folder join + state set as [`Self::active_org_shares_for_folder`], but carries the local
+    /// row id + item id for revocation.
+    pub fn active_org_share_ids_for_folder(
+        &self,
+        folder_id: &str,
+    ) -> Result<Vec<(String, Option<String>, String)>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT s.id, s.item_id, s.title
+                   FROM org_shares s
+                   LEFT JOIN notes n     ON n.meeting_id = s.meeting_id
+                   LEFT JOIN documents d ON d.id = s.document_id
+                  WHERE (s.state IN ('queued','uploaded','revoke_pending')
+                     OR (s.state = 'failed' AND s.item_id IS NOT NULL))
+                    AND (n.folder_id = ?1 OR d.folder_id = ?1)",
+            )
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![folder_id], |r| {
+                Ok((
+                    r.get::<_, String>(0)?,
+                    r.get::<_, Option<String>>(1)?,
+                    r.get::<_, Option<String>>(2)?.unwrap_or_default(),
+                ))
+            })
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    // The org partition is a decrypted REPLICA of the org feed, living in the dedicated `org_*`
+    // tables OUTSIDE the folder-lock domain (spec §"Trust model": org items are deliberately
+    // org-disclosed content — no folder seal/gate applies; SQLCipher protects them at rest). All
+    // writes here happen after the caller OPENED the OCK-sealed envelope (`share::org_envelope`),
+    // so the plaintext title/markdown are already the member's to see. NO PII in logs (ids/counts).
+
+    /// UPSERT one decrypted org feed item + (re)index its chunks. Idempotent on `item_id`: a re-pull
+    /// of the same seq REPLACES the row and re-chunks (clean replace via `index_org_item_chunks`). A
+    /// bumped `rev` (an update-share) overwrites the markdown + re-indexes. `content_sha256` is the
+    /// PLAINTEXT hash (the self-share dedup key). `embedder` is the member's OWN active embedder —
+    /// `None`/StubEmbedder ⇒ FTS-only (no int8 vectors written; the sync report flags `ftsOnly`).
+    ///
+    /// `author_user_id` (2026-07-15 root-cause fix, replaces the separate `set_org_item_author`
+    /// follow-up call as the ONLY writer of this column): the server-authoritative author id, when
+    /// the caller already knows it at upsert time — feed-ingest passes the feed entry's own
+    /// `author_user_id`; a share-time/republish-time local-replica upsert passes the CURRENT
+    /// session's own server user id (the caller IS the author in both those paths). `None` when the
+    /// caller genuinely doesn't know it (a light re-upsert, or a legacy call site). The
+    /// `ON CONFLICT` clause uses `COALESCE(excluded.author_user_id, org_items.author_user_id)` so a
+    /// `None` re-upsert can NEVER clobber an already-stamped author back to NULL — only a `Some`
+    /// value ever overwrites a previous value (and only with a fresher one, since every caller here
+    /// passes the authoritative id it has). This makes new/republished rows correct from the moment
+    /// they're born, with zero dependency on the `backfill_null_org_item_authors` self-heal (which
+    /// stays in place as a safety net for rows that predate this fix).
+    #[allow(clippy::too_many_arguments)]
+    pub fn upsert_org_item(
+        &self,
+        item_id: &str,
+        org_id: &str,
+        seq: u64,
+        author_hint: &str,
+        title: &str,
+        markdown: &str,
+        created_at: &str,
+        rev: u32,
+        generation: u32,
+        content_sha256: &[u8],
+        source_kind: Option<&str>,
+        author_user_id: Option<&str>,
+        embedder: Option<&dyn Embedder>,
+    ) -> Result<()> {
+        // Chunk + embed OUTSIDE the write lock (embedding is CPU/Metal work). The header carries the
+        // item title as provenance; the date axis is the item's created_at.
+        let chunks = crate::embed::chunk_note(title, created_at, markdown);
+        let vectors = match embedder {
+            Some(e) if !chunks.is_empty() => Some(e.embed_passage(&chunks)?),
+            _ => None, // model absent → FTS-only (int8 vectors come later on a re-embed).
+        };
+
+        let mut conn = self.lock();
+        let tx = conn.transaction().map_err(map_err)?;
+        // Replace the item row (idempotent upsert). CASCADE + the explicit vec purge below clear the
+        // old chunks first so a re-pull never leaves stale chunks/vectors.
+        Self::purge_org_item_chunks_tx(&tx, item_id)?;
+        tx.execute(
+            "INSERT INTO org_items
+               (item_id, org_id, seq, author_hint, title, markdown, created_at, rev, generation,
+                content_sha256, source_kind, author_user_id, tombstoned)
+             VALUES (?1,?2,?3,?4,?5,?6,?7,?8,?9,?10,?11,?12,0)
+             ON CONFLICT(item_id) DO UPDATE SET
+               org_id=excluded.org_id, seq=excluded.seq, author_hint=excluded.author_hint,
+               title=excluded.title, markdown=excluded.markdown, created_at=excluded.created_at,
+               rev=excluded.rev, generation=excluded.generation,
+               content_sha256=excluded.content_sha256, source_kind=excluded.source_kind,
+               author_user_id=COALESCE(excluded.author_user_id, org_items.author_user_id),
+               tombstoned=0",
+            rusqlite::params![
+                item_id,
+                org_id,
+                seq as i64,
+                author_hint,
+                title,
+                markdown,
+                created_at,
+                rev as i64,
+                generation as i64,
+                content_sha256,
+                source_kind,
+                author_user_id,
+            ],
+        )
+        .map_err(map_err)?;
+        {
+            let mut ins_chunk = tx
+                .prepare("INSERT INTO org_chunks (item_id, chunk_idx, text) VALUES (?1, ?2, ?3)")
+                .map_err(map_err)?;
+            // int8 vec0 → the value MUST be wrapped `vec_int8(?)` (the scale-spike partition format).
+            let mut ins_vec = tx
+                .prepare("INSERT INTO org_vec_chunks(chunk_id, embedding) VALUES (?1, vec_int8(?2))")
+                .map_err(map_err)?;
+            for (idx, text) in chunks.iter().enumerate() {
+                ins_chunk
+                    .execute(rusqlite::params![item_id, idx as i64, text])
+                    .map_err(map_err)?;
+                if let Some(vecs) = &vectors {
+                    if let Some(vector) = vecs.get(idx) {
+                        let chunk_id = tx.last_insert_rowid();
+                        let blob = crate::embed::vec_to_int8_blob(vector);
+                        ins_vec
+                            .execute(rusqlite::params![chunk_id, blob])
+                            .map_err(map_err)?;
+                    }
+                }
+            }
+        }
+        tx.commit().map_err(map_err)?;
+        Ok(())
+    }
+
+    /// TOMBSTONE an org item: mark the row `tombstoned=1` and DROP its chunks/vectors/FTS so the item
+    /// vanishes from retrieval, while the tombstone row keeps a re-pull idempotent (a later feed entry
+    /// for the same id is a no-op). Idempotent — tombstoning an unknown/already-tombstoned id is fine.
+    pub fn tombstone_org_item(&self, item_id: &str) -> Result<()> {
+        let mut conn = self.lock();
+        let tx = conn.transaction().map_err(map_err)?;
+        Self::purge_org_item_chunks_tx(&tx, item_id)?;
+        tx.execute(
+            "UPDATE org_items SET tombstoned = 1, markdown = '', title = '' WHERE item_id = ?1",
+            rusqlite::params![item_id],
+        )
+        .map_err(map_err)?;
+        tx.commit().map_err(map_err)?;
+        Ok(())
+    }
+
+    /// LEAVE-A-ORG CONSENT PURGE: drop the ENTIRE decrypted replica of one org — every `org_items`
+    /// row plus its derived `org_chunks` / `org_vec_chunks` / `fts_org_chunks` tokens — in ONE atomic
+    /// tx. Called by `org_leave` so a departed member keeps NO searchable copy of colleagues' shared
+    /// content (leak/consent invariant): the OCK cache is dropped and `org_state` deleted at the
+    /// command layer, but WITHOUT this the plaintext replica lingered forever and `org_search` could
+    /// still return it. Order: vec0 first (its FK-less rowid mirrors `org_chunks.id`), then
+    /// `org_chunks` (whose DELETE fires the `fts_org_chunks_ad` trigger, purging the keyword tokens),
+    /// then the `org_items` header rows. Idempotent; an unknown org id is a no-op. Content-free log
+    /// (org id + counts, never titles/bodies).
+    pub fn purge_org_replica(&self, org_id: &str) -> Result<()> {
+        let mut conn = self.lock();
+        let tx = conn.transaction().map_err(map_err)?;
+        // vec0 KNN rows for every chunk of every item in this org (the FK-less mirror table).
+        tx.execute(
+            "DELETE FROM org_vec_chunks WHERE chunk_id IN
+               (SELECT oc.id FROM org_chunks oc
+                  JOIN org_items oi ON oi.item_id = oc.item_id
+                 WHERE oi.org_id = ?1)",
+            rusqlite::params![org_id],
+        )
+        .map_err(map_err)?;
+        // Source chunks (their DELETE fires the FTS `_ad` trigger → keyword tokens purged).
+        tx.execute(
+            "DELETE FROM org_chunks WHERE item_id IN
+               (SELECT item_id FROM org_items WHERE org_id = ?1)",
+            rusqlite::params![org_id],
+        )
+        .map_err(map_err)?;
+        // Finally the item headers (the decrypted markdown/title replica).
+        let items = tx
+            .execute(
+                "DELETE FROM org_items WHERE org_id = ?1",
+                rusqlite::params![org_id],
+            )
+            .map_err(map_err)?;
+        tx.commit().map_err(map_err)?;
+        tracing::info!(target: "org", items, "purged org replica on leave");
+        Ok(())
+    }
+
+    /// Delete an org item's `org_chunks` + `org_vec_chunks` rows within an EXISTING tx. vec0 first
+    /// (its FK-less rowid mirrors `org_chunks.id`), then the source rows (whose DELETE fires the FTS
+    /// `_ad` trigger, purging the tokens). Mirrors [`Db::purge_doc_chunks_tx`].
+    fn purge_org_item_chunks_tx(tx: &rusqlite::Transaction<'_>, item_id: &str) -> Result<()> {
+        tx.execute(
+            "DELETE FROM org_vec_chunks WHERE chunk_id IN
+               (SELECT id FROM org_chunks WHERE item_id = ?1)",
+            rusqlite::params![item_id],
+        )
+        .map_err(map_err)?;
+        tx.execute(
+            "DELETE FROM org_chunks WHERE item_id = ?1",
+            rusqlite::params![item_id],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// The synced feed cursor (`org_state.last_seq`) for an org — the max `seq` ingested so far.
+    /// Returns 0 when the org is unknown/never synced. (Companion to Core's monotonic
+    /// [`Db::set_org_last_seq`].)
+    pub fn org_last_seq_for(&self, org_id: &str) -> Result<u64> {
+        Ok(self
+            .get_org_state(org_id)?
+            .map(|s| s.last_seq.max(0) as u64)
+            .unwrap_or(0))
+    }
+
+    /// GATED-FREE (no folder lock applies to org items) semantic KNN over the int8 org partition:
+    /// the top-`k` nearest `org_vec_chunks` for the int8-quantized `query_vec`, joined to their
+    /// (non-tombstoned) items, deduped to one hit per item (nearest). `query_vec` is the member's OWN
+    /// f32 query embedding — it is int8-quantized here so it is comparable to the stored int8 vectors.
+    ///
+    /// PER-INSTANCE ORG TOGGLE: joined to `org_state` and filtered on `context_enabled = 1` — a
+    /// disabled org's chunks are EXCLUDED at the SQL level, never read into Rust at all. This is the
+    /// hard data-level gate (not a UI hide): a caller cannot accidentally surface a disabled org's
+    /// content by forgetting to filter it downstream.
+    pub fn search_org_chunks_knn(&self, query_vec: &[f32], k: i64) -> Result<Vec<OrgChunkHit>> {
+        if query_vec.is_empty() || k <= 0 {
+            return Ok(Vec::new());
+        }
+        let conn = self.lock();
+        // KNN isolated to the vec0 table in a CTE (a vec0 query allows a single MATCH+k); the item
+        // columns + the tombstone/context-enabled filters are joined OUTSIDE it.
+        let sql = "WITH knn(chunk_id, distance) AS (
+                 SELECT chunk_id, distance FROM org_vec_chunks
+                  WHERE embedding MATCH vec_int8(?1) AND k = ?2
+                  ORDER BY distance
+             )
+             SELECT oi.item_id, oi.author_hint, oi.title, oc.text, oi.content_sha256, knn.distance
+               FROM knn
+               JOIN org_chunks oc ON oc.id = knn.chunk_id
+               JOIN org_items oi ON oi.item_id = oc.item_id
+               JOIN org_state os ON os.org_id = oi.org_id
+              WHERE oi.tombstoned = 0 AND os.context_enabled = 1
+              ORDER BY knn.distance ASC, oi.item_id ASC";
+        let blob = crate::embed::vec_to_int8_blob(query_vec);
+        let mut stmt = conn.prepare(sql).map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![blob, k], |row| {
+                Ok(OrgChunkHit {
+                    item_id: row.get(0)?,
+                    author_hint: row.get(1)?,
+                    title: row.get(2)?,
+                    snippet: row.get(3)?,
+                    content_sha256: row.get::<_, Option<Vec<u8>>>(4)?.unwrap_or_default(),
+                })
+            })
+            .map_err(map_err)?;
+        Self::dedup_org_hits_by_item(rows)
+    }
+
+    /// KEYWORD (FTS5/BM25) leg over the org partition — the model-free twin of
+    /// [`Db::search_org_chunks_knn`], so org text is reachable on a DEFAULT install (no e5 model).
+    /// Same `context_enabled = 1` per-instance org filter as the KNN leg — see its doc.
+    ///
+    /// CRITICAL (scale-spike finding #2): the `LIMIT` is PUSHED DOWN into the SQL (bm25-ordered),
+    /// NOT applied in Rust after reading every match — the unbounded production reader hit an 8.8 s
+    /// p95 tail at 1M chunks. We over-fetch a small multiple of `limit` (so per-item dedup still has
+    /// candidates) then cap in Rust; the SQL ceiling is the real bound.
+    pub fn search_org_chunks_fts(&self, query: &str, limit: i64) -> Result<Vec<OrgChunkHit>> {
+        if limit <= 0 {
+            return Ok(Vec::new());
+        }
+        let Some(match_expr) = fts_match_query(query.trim()) else {
+            return Ok(Vec::new()); // punctuation-only / empty query → no hits, never an FTS error.
+        };
+        let conn = self.lock();
+        // Over-fetch a bounded multiple of `limit` so per-item dedup has candidates, but keep the SQL
+        // LIMIT as the hard bound (spike #2). 8× is generous for a per-item dedup at small `limit`.
+        let sql_cap = limit.saturating_mul(8).clamp(limit, 512);
+        let sql = "SELECT oi.item_id, oi.author_hint, oi.title, oc.text, oi.content_sha256,
+                          bm25(fts_org_chunks) AS rank
+               FROM fts_org_chunks
+               JOIN org_chunks oc ON oc.id = fts_org_chunks.rowid
+               JOIN org_items oi ON oi.item_id = oc.item_id
+               JOIN org_state os ON os.org_id = oi.org_id
+              WHERE fts_org_chunks MATCH ?1 AND oi.tombstoned = 0 AND os.context_enabled = 1
+              ORDER BY rank ASC, oi.item_id ASC
+              LIMIT ?2";
+        let mut stmt = conn.prepare(sql).map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![match_expr, sql_cap], |row| {
+                Ok(OrgChunkHit {
+                    item_id: row.get(0)?,
+                    author_hint: row.get(1)?,
+                    title: row.get(2)?,
+                    snippet: row.get(3)?,
+                    content_sha256: row.get::<_, Option<Vec<u8>>>(4)?.unwrap_or_default(),
+                })
+            })
+            .map_err(map_err)?;
+        let mut hits = Self::dedup_org_hits_by_item(rows)?;
+        hits.truncate(limit as usize);
+        Ok(hits)
+    }
+
+    /// Dedup a stream of org chunk hits to ONE per item (first-seen = best-ranked, since callers
+    /// order by distance/bm25 ascending). Shared by both retrieval legs.
+    fn dedup_org_hits_by_item(
+        rows: impl Iterator<Item = rusqlite::Result<OrgChunkHit>>,
+    ) -> Result<Vec<OrgChunkHit>> {
+        let mut seen: HashSet<String> = HashSet::new();
+        let mut hits = Vec::new();
+        for r in rows {
+            let hit = r.map_err(map_err)?;
+            if !seen.insert(hit.item_id.clone()) {
+                continue;
+            }
+            hits.push(hit);
+        }
+        Ok(hits)
+    }
+
+    /// The full decrypted org item (for the read-only FE viewer). `None` for an unknown, TOMBSTONED,
+    /// OR per-instance-DISABLED item's org (a stale citation/bookmark to `/org-item/:id` must not read
+    /// through the toggle — same `context_enabled = 1` gate as `search_org_chunks_knn`/`_fts`). No lock
+    /// gate otherwise — org items are deliberately org-disclosed content.
+    pub fn get_org_item(&self, item_id: &str) -> Result<Option<crate::storage::models::OrgItemDetail>> {
+        let conn = self.lock();
+        conn.query_row(
+            "SELECT oi.item_id, oi.author_hint, oi.title, oi.created_at, oi.rev, oi.markdown
+               FROM org_items oi
+               JOIN org_state os ON os.org_id = oi.org_id
+              WHERE oi.item_id = ?1 AND oi.tombstoned = 0 AND os.context_enabled = 1",
+            rusqlite::params![item_id],
+            |r| {
+                Ok(crate::storage::models::OrgItemDetail {
+                    item_id: r.get(0)?,
+                    author_hint: r.get(1)?,
+                    title: r.get(2)?,
+                    created_at: r.get(3)?,
+                    rev: r.get::<_, i64>(4)? as u32,
+                    markdown: r.get(5)?,
+                    // The DB layer has no session context — the `org_get_item` command computes the real
+                    // value by comparing the stored `author_user_id` with the caller's `server_user_id`.
+                    editable: false,
+                })
+            },
+        )
+        .optional()
+        .map_err(map_err)
+    }
+
+    /// Stamp an org item's `author_user_id` (the server account id of its author, from the feed). Called
+    /// right after `upsert_org_item` at feed-ingest so a second machine can recognise its OWN items and
+    /// offer edit-in-place. Idempotent; a no-op for an unknown id. (2026-07-14.)
+    pub fn set_org_item_author(&self, item_id: &str, author_user_id: &str) -> Result<()> {
+        let conn = self.lock();
+        conn.execute(
+            "UPDATE org_items SET author_user_id = ?2 WHERE item_id = ?1",
+            rusqlite::params![item_id, author_user_id],
+        )
+        .map_err(map_err)?;
+        Ok(())
+    }
+
+    /// The item ids of this org's LIVE (non-tombstoned) local replica rows that are still missing
+    /// `author_user_id` — the stale-ingest gap (rows ingested before the column/stamping existed, or
+    /// via the local-replica upsert at share/republish time, whose cursor has already advanced past
+    /// them so a normal cursor-based feed pull never re-visits them). Used by the sync-tick backfill
+    /// (`backfill_null_org_item_authors`) to know whether a full-feed re-pull is worth doing at all —
+    /// an empty result short-circuits the backfill on every ordinary sync once a device has caught up.
+    /// (2026-07-15.)
+    pub fn org_item_ids_with_null_author(&self, org_id: &str) -> Result<Vec<String>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT item_id FROM org_items
+                   WHERE org_id = ?1 AND tombstoned = 0 AND author_user_id IS NULL",
+            )
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![org_id], |r| r.get::<_, String>(0))
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// The context the `org_update_own_item` egress command needs to re-publish an edited org item the
+    /// caller authored (org id, current rev, original created_at + source_kind, stored author id). `None`
+    /// for an unknown / tombstoned item. (2026-07-14.)
+    pub fn org_item_edit_ctx(
+        &self,
+        item_id: &str,
+    ) -> Result<Option<crate::storage::models::OrgItemEditCtx>> {
+        let conn = self.lock();
+        conn.query_row(
+            "SELECT org_id, rev, created_at, source_kind, author_user_id
+               FROM org_items WHERE item_id = ?1 AND tombstoned = 0",
+            rusqlite::params![item_id],
+            |r| {
+                Ok(crate::storage::models::OrgItemEditCtx {
+                    org_id: r.get(0)?,
+                    rev: r.get::<_, i64>(1)? as u32,
+                    created_at: r.get(2)?,
+                    source_kind: r.get::<_, Option<String>>(3)?,
+                    author_user_id: r.get::<_, Option<String>>(4)?,
+                })
+            },
+        )
+        .optional()
+        .map_err(map_err)
+    }
+
+    /// The browsable LIST of one org's live (non-tombstoned) items — headers only (no `markdown`
+    /// body; that's [`Db::get_org_item`]). Newest-first by feed `seq`. This is what lets a member SEE
+    /// what colleagues shared into the org instead of only search-hitting it. Org items are
+    /// deliberately org-disclosed content (no folder lock gate applies); the COMMAND layer re-checks
+    /// the caller is a local member of `org_id` before calling this.
+    ///
+    /// `kind` is now populated DIRECTLY from the stored `source_kind` column (opened off the item's
+    /// `OrgEnvelope` at ingest — see `upsert_org_item`) for EVERY item, not just ones this device
+    /// published: a v2-envelope item from a colleague now classifies correctly. Stays `None` for a row
+    /// ingested before this column existed, or from a peer still on an old v1-only client (honest
+    /// "unclassified", never guessed). `list_org_items_inner` may still override this with the
+    /// owned-item resolver for the caller's OWN items (correct even when the column is somehow null).
+    pub fn list_org_items(
+        &self,
+        org_id: &str,
+    ) -> Result<Vec<crate::storage::models::OrgItemHeader>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT item_id, title, author_hint, created_at, seq, source_kind
+                   FROM org_items
+                  WHERE org_id = ?1 AND tombstoned = 0
+                  ORDER BY seq DESC",
+            )
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![org_id], |r| {
+                Ok(crate::storage::models::OrgItemHeader {
+                    item_id: r.get(0)?,
+                    title: r.get(1)?,
+                    author_hint: r.get(2)?,
+                    created_at: r.get(3)?,
+                    seq: r.get::<_, i64>(4)? as u64,
+                    // Direct from storage now (see doc comment above); `list_org_items_inner` may still
+                    // enrich/override for the caller's own items via the local `org_shares` resolver.
+                    kind: r.get::<_, Option<String>>(5)?,
+                    owned_source: None,
+                })
+            })
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// COUNT of one org's live (non-tombstoned) RECEIVED items — the size of the local org replica
+    /// (what colleagues shared IN). Distinct from the outbound `org_shares` count (what THIS member
+    /// published OUT); the two were conflated so the Settings item count showed the caller's own
+    /// uploads and read "0 items" to a receiver. Content-free.
+    ///
+    /// PER-INSTANCE ORG TOGGLE: joined to `org_state` and filtered on `context_enabled = 1` — same
+    /// gate as `search_org_chunks_knn`/`_fts`/`get_org_item`/`list_org_items_inner`. Without this a
+    /// disabled org's `received_count` stayed stale/inflated (the raw local-replica row count) even
+    /// though every other read of the same table (search, browse) correctly reports it as empty —
+    /// the count and the actual gated content must agree for the same org.
+    pub fn count_org_items(&self, org_id: &str) -> Result<u32> {
+        let conn = self.lock();
+        conn.query_row(
+            "SELECT COUNT(*)
+               FROM org_items oi
+               JOIN org_state os ON os.org_id = oi.org_id
+              WHERE oi.org_id = ?1 AND oi.tombstoned = 0 AND os.context_enabled = 1",
+            rusqlite::params![org_id],
+            |r| r.get::<_, i64>(0),
+        )
+        .map(|n| n as u32)
+        .map_err(map_err)
+    }
+
+    /// Every non-null `content_sha256` from the local `org_shares` rows (across all orgs the user has
+    /// shared into) — the SELF-SHARE dedup key set. A retrieval hit whose hash is in this set is the
+    /// caller's OWN published item and is relabelled/dropped so a member never sees their own share
+    /// echoed back as an "org" result. Content-free (opaque hashes only).
+    pub fn all_org_shared_content_hashes(&self) -> Result<Vec<Vec<u8>>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare("SELECT content_sha256 FROM org_shares WHERE content_sha256 IS NOT NULL")
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map([], |r| r.get::<_, Vec<u8>>(0))
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+
+    /// LIVE org item ids that still LACK any int8 vector (chunks present but no `org_vec_chunks`) —
+    /// the re-embed backlog once a real embedder appears on a member that ingested FTS-only. Bounded
+    /// by `limit`. Empty when every live item is already embedded (or FTS-only with no chunks).
+    pub fn org_items_needing_embed(&self, org_id: &str, limit: i64) -> Result<Vec<String>> {
+        let conn = self.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT DISTINCT oc.item_id
+                   FROM org_chunks oc
+                   JOIN org_items oi ON oi.item_id = oc.item_id
+                  WHERE oi.org_id = ?1 AND oi.tombstoned = 0
+                    AND NOT EXISTS (SELECT 1 FROM org_vec_chunks v WHERE v.chunk_id = oc.id)
+                  LIMIT ?2",
+            )
+            .map_err(map_err)?;
+        let rows = stmt
+            .query_map(rusqlite::params![org_id, limit], |r| r.get::<_, String>(0))
+            .map_err(map_err)?;
+        let mut out = Vec::new();
+        for r in rows {
+            out.push(r.map_err(map_err)?);
+        }
+        Ok(out)
+    }
+}
